Use the new single replacingmergetree task events table for replication
