+++ apps/webapp/app/services/runsDashboardService.server.ts
@@ -1,10 +1,9 @@
 import type { ClickHouse } from "@internal/clickhouse";
-import { TaskRunError } from "@trigger.dev/core/v3/schemas";
-import { RuntimeEnvironmentType, TaskRun, TaskRunStatus } from "@trigger.dev/database";
-import { logger } from "./logger.server";
-import { EventEmitter } from "node:events";
-import { parsePacket } from "@trigger.dev/core/v3/utils/ioSerialization";
 import { EventBusEvents } from "@internal/run-engine";
+import { parsePacket } from "@trigger.dev/core/v3/utils/ioSerialization";
+import { RuntimeEnvironmentType, TaskRun, TaskRunStatus } from "@trigger.dev/database";
+import { EventEmitter } from "node:events";
+import { logger } from "./logger.server";
 
 export class RunsDashboardService {
   constructor(private readonly clickhouse: ClickHouse) {}
@@ -14,290 +13,47 @@ export class RunsDashboardService {
   });
 
   async runAttemptStarted(event: RunDashboardEventRunAttemptStarted) {
-    const [insertError, insertResult] = await this.clickhouse.runEvents.insert({
-      environment_id: event.environment.id,
-      organization_id: event.organization.id,
-      project_id: event.project.id,
-      run_id: event.run.id,
-      status: event.run.status,
-      attempt: event.run.attemptNumber ?? 1,
-      event_time: event.time.getTime(),
-      updated_at: event.run.updatedAt.getTime(),
-      created_at: event.run.createdAt.getTime(),
-      base_cost_in_cents: event.run.baseCostInCents,
-      executed_at: event.run.executedAt ? event.run.executedAt.getTime() : undefined,
-      event_name: "attempt_started",
-    });
-
-    if (insertError) {
-      this.logger.error("RunsDashboardService: runAttemptStarted", {
-        error: insertError,
-        event,
-      });
-    }
-
-    return insertResult?.executed === true;
+    // Noop for now
   }
 
   async runEnqueuedAfterDelay(event: RunDashboardEventRunEnqueuedAfterDelay) {
-    const [insertError, insertResult] = await this.clickhouse.runEvents.insert({
-      environment_id: event.environment.id,
-      organization_id: event.organization.id,
-      project_id: event.project.id,
-      run_id: event.run.id,
-      status: event.run.status,
-      event_time: event.time.getTime(),
-      updated_at: event.run.updatedAt.getTime(),
-      created_at: event.run.createdAt.getTime(),
-      event_name: "enqueued_after_delay",
-    });
-
-    if (insertError) {
-      this.logger.error("RunsDashboardService: runEnqueuedAfterDelay", {
-        error: insertError,
-        event,
-      });
-    }
-
-    return insertResult?.executed === true;
+    // Noop for now
   }
 
   async runDelayRescheduled(event: RunDashboardEventRunDelayRescheduled) {
-    const [insertError, insertResult] = await this.clickhouse.runEvents.insert({
-      environment_id: event.environment.id,
-      organization_id: event.organization.id,
-      project_id: event.project.id,
-      run_id: event.run.id,
-      status: event.run.status,
-      event_time: event.time.getTime(),
-      updated_at: event.run.updatedAt.getTime(),
-      created_at: event.run.createdAt.getTime(),
-      delay_until: event.run.delayUntil ? event.run.delayUntil.getTime() : undefined,
-      event_name: "delay_rescheduled",
-    });
-
-    if (insertError) {
-      this.logger.error("RunsDashboardService: runDelayRescheduled", {
-        error: insertError,
-        event,
-      });
-    }
-
-    return insertResult?.executed === true;
+    // Noop for now
   }
 
   async runLocked(event: RunDashboardEventRunLocked) {
-    const [insertError, insertResult] = await this.clickhouse.runEvents.insert({
-      environment_id: event.environment.id,
-      organization_id: event.organization.id,
-      project_id: event.project.id,
-      run_id: event.run.id,
-      status: event.run.status,
-      event_time: event.time.getTime(),
-      updated_at: event.run.updatedAt.getTime(),
-      created_at: event.run.createdAt.getTime(),
-      base_cost_in_cents: event.run.baseCostInCents,
-      task_version: event.run.taskVersion ?? undefined,
-      sdk_version: event.run.sdkVersion ?? undefined,
-      cli_version: event.run.cliVersion ?? undefined,
-      machine_preset: event.run.machinePreset ?? undefined,
-      executed_at: event.run.startedAt ? event.run.startedAt.getTime() : undefined,
-      event_name: "locked",
-    });
-
-    if (insertError) {
-      this.logger.error("RunsDashboardService: runLocked", {
-        error: insertError,
-        event,
-      });
-    }
-
-    return insertResult?.executed === true;
+    // Noop for now
   }
 
   async runStatusChanged(event: RunDashboardEventRunStatusChanged) {
-    if (!event.organization.id || !event.project.id || !event.environment.id) {
-      return false;
-    }
-
-    const [insertError, insertResult] = await this.clickhouse.runEvents.insert({
-      environment_id: event.environment.id,
-      organization_id: event.organization.id,
-      project_id: event.project.id,
-      run_id: event.run.id,
-      status: event.run.status,
-      event_time: event.time.getTime(),
-      updated_at: event.run.updatedAt.getTime(),
-      created_at: event.run.createdAt.getTime(),
-      event_name: "status_changed",
-    });
-
-    if (insertError) {
-      this.logger.error("RunsDashboardService: runStatusChanged", {
-        error: insertError,
-        event,
-      });
-    }
-
-    return insertResult?.executed === true;
+    // Noop for now
   }
 
   async runExpired(event: RunDashboardEventRunExpired) {
-    const [insertError, insertResult] = await this.clickhouse.runEvents.insert({
-      environment_id: event.environment.id,
-      organization_id: event.organization.id,
-      project_id: event.project.id,
-      run_id: event.run.id,
-      status: event.run.status,
-      event_time: event.time.getTime(),
-      updated_at: event.run.updatedAt.getTime(),
-      created_at: event.run.createdAt.getTime(),
-      expired_at: event.run.expiredAt ? event.run.expiredAt.getTime() : undefined,
-      event_name: "run_expired",
-    });
-
-    if (insertError) {
-      this.logger.error("RunsDashboardService: runExpired", {
-        error: insertError,
-        event,
-      });
-    }
-
-    return insertResult?.executed === true;
+    // Noop for now
   }
 
   async runSucceeded(event: RunDashboardEventRunSucceeded) {
-    const output = await this.#prepareOutput(event.run);
-
-    const [insertError, insertResult] = await this.clickhouse.runEvents.insert({
-      environment_id: event.environment.id,
-      organization_id: event.organization.id,
-      project_id: event.project.id,
-      run_id: event.run.id,
-      status: event.run.status,
-      event_time: event.time.getTime(),
-      updated_at: event.run.updatedAt.getTime(),
-      created_at: event.run.createdAt.getTime(),
-      completed_at: event.run.completedAt ? event.run.completedAt.getTime() : undefined,
-      usage_duration_ms: event.run.usageDurationMs,
-      cost_in_cents: event.run.costInCents,
-      output: output,
-      attempt: event.run.attemptNumber,
-      event_name: "succeeded",
-    });
-
-    if (insertError) {
-      this.logger.error("RunsDashboardService: runSucceeded", {
-        error: insertError,
-        event,
-      });
-    }
-
-    return insertResult?.executed === true;
+    // Noop for now
   }
 
   async runFailed(event: RunDashboardEventRunFailed) {
-    const [insertError, insertResult] = await this.clickhouse.runEvents.insert({
-      environment_id: event.environment.id,
-      organization_id: event.organization.id,
-      project_id: event.project.id,
-      run_id: event.run.id,
-      status: event.run.status,
-      event_time: event.time.getTime(),
-      updated_at: event.run.updatedAt.getTime(),
-      created_at: event.run.createdAt.getTime(),
-      completed_at: event.run.completedAt ? event.run.completedAt.getTime() : undefined,
-      error: event.run.error,
-      attempt: event.run.attemptNumber,
-      usage_duration_ms: event.run.usageDurationMs,
-      cost_in_cents: event.run.costInCents,
-      event_name: "failed",
-    });
-
-    if (insertError) {
-      this.logger.error("RunsDashboardService: runFailed", {
-        error: insertError,
-        event,
-      });
-    }
-
-    return insertResult?.executed === true;
+    // Noop for now
   }
 
   async runRetryScheduled(event: RunDashboardEventRunRetryScheduled) {
-    const [insertError, insertResult] = await this.clickhouse.runEvents.insert({
-      environment_id: event.environment.id,
-      organization_id: event.organization.id,
-      project_id: event.environment.projectId,
-      run_id: event.run.id,
-      status: event.run.status,
-      event_time: event.time.getTime(),
-      updated_at: event.run.updatedAt.getTime(),
-      created_at: event.run.createdAt.getTime(),
-      machine_preset: event.run.nextMachineAfterOOM ?? undefined,
-      attempt: event.run.attemptNumber,
-      error: event.run.error,
-      event_name: "retry_scheduled",
-    });
-
-    if (insertError) {
-      this.logger.error("RunsDashboardService: runRetryScheduled", {
-        error: insertError,
-        event,
-      });
-    }
-
-    return insertResult?.executed === true;
+    // Noop for now
   }
 
   async runCancelled(event: RunDashboardEventRunCancelled) {
-    const [insertError, insertResult] = await this.clickhouse.runEvents.insert({
-      environment_id: event.environment.id,
-      organization_id: event.organization.id,
-      project_id: event.project.id,
-      run_id: event.run.id,
-      status: event.run.status,
-      event_time: event.time.getTime(),
-      updated_at: event.run.updatedAt.getTime(),
-      created_at: event.run.createdAt.getTime(),
-      completed_at: event.run.completedAt ? event.run.completedAt.getTime() : undefined,
-      error: event.run.error ? (event.run.error as TaskRunError) : undefined,
-      attempt: event.run.attemptNumber,
-      event_name: "cancelled",
-    });
-
-    if (insertError) {
-      this.logger.error("RunsDashboardService: runCancelled", {
-        error: insertError,
-        event,
-      });
-    }
-
-    return insertResult?.executed === true;
+    // Noop for now
   }
 
   async runTagsUpdated(event: RunDashboardEventRunTagsUpdated) {
-    const [insertError, insertResult] = await this.clickhouse.runEvents.insert({
-      environment_id: event.environment.id,
-      organization_id: event.organization.id,
-      project_id: event.project.id,
-      run_id: event.run.id,
-      status: event.run.status,
-      event_time: event.time.getTime(),
-      updated_at: event.run.updatedAt.getTime(),
-      created_at: event.run.createdAt.getTime(),
-      tags: event.run.tags,
-      event_name: "tags_updated",
-    });
-
-    if (insertError) {
-      this.logger.error("RunsDashboardService: runTagsUpdated", {
-        error: insertError,
-        event,
-      });
-    }
-
-    return insertResult?.executed === true;
+    // Noop for now
   }
 
   async runCreated(
@@ -306,65 +62,7 @@ export class RunsDashboardService {
     environmentType: RuntimeEnvironmentType,
     organizationId: string
   ) {
-    const payload = await this.#preparePayload(taskRun);
-
-    const [insertError, insertResult] = await this.clickhouse.runEvents.insert({
-      environment_id: taskRun.runtimeEnvironmentId,
-      environment_type: environmentType,
-      organization_id: organizationId,
-      project_id: taskRun.projectId,
-      run_id: taskRun.id,
-      friendly_id: taskRun.friendlyId,
-      attempt: taskRun.attemptNumber ?? 1,
-      engine: taskRun.engine,
-      status: taskRun.status,
-      task_identifier: taskRun.taskIdentifier,
-      queue: taskRun.queue,
-      schedule_id: taskRun.scheduleId ?? undefined,
-      batch_id: taskRun.batchId ?? undefined,
-      event_time: eventTime.getTime(),
-      created_at: taskRun.createdAt.getTime(),
-      updated_at: taskRun.updatedAt.getTime(),
-      completed_at: taskRun.completedAt ? taskRun.completedAt.getTime() : undefined,
-      started_at: taskRun.startedAt ? taskRun.startedAt.getTime() : undefined,
-      executed_at: taskRun.executedAt ? taskRun.executedAt.getTime() : undefined,
-      delay_until: taskRun.delayUntil ? taskRun.delayUntil.getTime() : undefined,
-      queued_at: taskRun.queuedAt ? taskRun.queuedAt.getTime() : undefined,
-      expired_at: taskRun.expiredAt ? taskRun.expiredAt.getTime() : undefined,
-      usage_duration_ms: taskRun.usageDurationMs,
-      tags: taskRun.runTags,
-      payload: payload,
-      task_version: taskRun.taskVersion ?? undefined,
-      sdk_version: taskRun.sdkVersion ?? undefined,
-      cli_version: taskRun.cliVersion ?? undefined,
-      machine_preset: taskRun.machinePreset ?? undefined,
-      is_test: taskRun.isTest ?? false,
-      root_run_id: taskRun.rootTaskRunId ?? undefined,
-      parent_run_id: taskRun.parentTaskRunId ?? undefined,
-      depth: taskRun.depth ?? 0,
-      span_id: taskRun.spanId ?? undefined,
-      trace_id: taskRun.traceId ?? undefined,
-      idempotency_key: taskRun.idempotencyKey ?? undefined,
-      expiration_ttl: taskRun.ttl ?? undefined,
-      cost_in_cents: taskRun.costInCents ?? undefined,
-      base_cost_in_cents: taskRun.baseCostInCents ?? undefined,
-      event_name: "created",
-    });
-
-    if (insertError) {
-      this.logger.error("RunsDashboardService: upsertRun", {
-        error: insertError,
-        taskRun,
-      });
-    } else {
-      this.logger.info("RunsDashboardService: upsertRun", {
-        id: taskRun.id,
-        friendlyId: taskRun.friendlyId,
-        status: taskRun.status,
-      });
-    }
-
-    return insertResult?.executed === true;
+    // Noop for now
   }
 
   async #preparePayload(run: TaskRun): Promise<unknown | undefined> {
@@ -399,26 +97,6 @@ export class RunsDashboardService {
     const packet = {
       data: run.output,
       dataType: run.outputType,
-    };
-
-    return await parsePacket(packet);
-  }
-
-  async #prepareMetadata(run: {
-    metadata: string | undefined;
-    metadataType: string;
-  }): Promise<unknown | undefined> {
-    if (!run.metadata) {
-      return undefined;
-    }
-
-    if (run.metadataType !== "application/json" && run.metadataType !== "application/super+json") {
-      return undefined;
-    }
-
-    const packet = {
-      data: run.metadata,
-      dataType: run.metadataType,
     };
 
     return await parsePacket(packet);
