--- internal-packages/clickhouse/src/runEvents.test.ts
@@ -1,830 +1,0 @@
-import { clickhouseTest } from "@internal/testcontainers";
-import { z } from "zod";
-import { ClickhouseClient } from "./client/client.js";
-import { insertRunEvents } from "./runEvents.js";
-
-describe("Run Events", () => {
-  clickhouseTest("should be able to insert run events", async ({ clickhouseContainer }) => {
-    const client = new ClickhouseClient({
-      name: "test",
-      url: clickhouseContainer.getConnectionUrl(),
-    });
-
-    const insert = insertRunEvents(client, {
-      async_insert: 0, // turn off async insert for this test
-    });
-
-    const [insertError, insertResult] = await insert([
-      {
-        environment_id: "env_1234",
-        event_name: "created",
-        environment_type: "DEVELOPMENT",
-        organization_id: "org_1234",
-        project_id: "project_1234",
-        run_id: "run_1234",
-        friendly_id: "friendly_1234",
-        attempt: 1,
-        engine: "V2",
-        status: "PENDING",
-        task_identifier: "my-task",
-        queue: "my-queue",
-        schedule_id: "schedule_1234",
-        batch_id: "batch_1234",
-        event_time: Date.now(),
-        created_at: Date.now(),
-        updated_at: Date.now(),
-        completed_at: undefined,
-        tags: ["tag1", "tag2"],
-        payload: {
-          key: "value",
-        },
-        output: {
-          key: "value",
-        },
-        error: {
-          type: "BUILT_IN_ERROR",
-          name: "Error",
-          message: "error",
-          stackTrace: "stack trace",
-        },
-        usage_duration_ms: 1000,
-        cost_in_cents: 100,
-        task_version: "1.0.0",
-        sdk_version: "1.0.0",
-        cli_version: "1.0.0",
-        machine_preset: "small-1x",
-        is_test: true,
-        span_id: "span_1234",
-        trace_id: "trace_1234",
-        idempotency_key: "idempotency_key_1234",
-        expiration_ttl: "1h",
-        root_run_id: "root_run_1234",
-        parent_run_id: "parent_run_1234",
-        depth: 1,
-      },
-    ]);
-
-    expect(insertError).toBeNull();
-    expect(insertResult).toEqual(expect.objectContaining({ executed: true }));
-    expect(insertResult?.summary?.written_rows).toEqual("2");
-
-    const query = client.query({
-      name: "query-run-events",
-      query: "SELECT * FROM trigger_dev.raw_run_events_v1",
-      schema: z.object({
-        environment_id: z.string(),
-        run_id: z.string(),
-      }),
-      params: z.object({
-        run_id: z.string(),
-      }),
-    });
-
-    const [queryError, result] = await query({ run_id: "run_1234" });
-
-    expect(queryError).toBeNull();
-    expect(result).toEqual(
-      expect.arrayContaining([
-        expect.objectContaining({
-          environment_id: "env_1234",
-          run_id: "run_1234",
-        }),
-      ])
-    );
-  });
-
-  clickhouseTest(
-    "should be able to handle multiple run events",
-    async ({ clickhouseContainer }) => {
-      const client = new ClickhouseClient({
-        name: "test",
-        url: clickhouseContainer.getConnectionUrl(),
-      });
-
-      const insert = insertRunEvents(client, {
-        async_insert: 0, // turn off async insert for this test
-      });
-
-      const [insertError, insertResult] = await insert([
-        {
-          environment_id: "cm9kddfcs01zqdy88ld9mmrli",
-          organization_id: "cm8zs78wb0002dy616dg75tv3",
-          project_id: "cm9kddfbz01zpdy88t9dstecu",
-          run_id: "cma45oli70002qrdy47w0j4n7",
-          event_name: "created",
-          environment_type: "PRODUCTION",
-          friendly_id: "run_cma45oli70002qrdy47w0j4n7",
-          attempt: 1,
-          engine: "V2",
-          status: "PENDING",
-          task_identifier: "retry-task",
-          queue: "task/retry-task",
-          schedule_id: null,
-          batch_id: null,
-          root_run_id: null,
-          parent_run_id: null,
-          depth: 0,
-          span_id: "538677637f937f54",
-          trace_id: "20a28486b0b9f50c647b35e8863e36a5",
-          idempotency_key: null,
-          event_time: new Date("2025-04-30 16:34:04.341").getTime(),
-          created_at: new Date("2025-04-30 16:34:04.312").getTime(),
-          updated_at: new Date("2025-04-30 16:34:04.312").getTime(),
-          started_at: null,
-          executed_at: null,
-          completed_at: null,
-          delay_until: null,
-          queued_at: new Date("2025-04-30 16:34:04.311").getTime(),
-          expired_at: null,
-          expiration_ttl: null,
-          usage_duration_ms: 0,
-          cost_in_cents: 0,
-          base_cost_in_cents: 0,
-          payload: { failCount: "3" },
-          output: null,
-          error: null,
-          tags: [],
-          task_version: null,
-          sdk_version: null,
-          cli_version: null,
-          machine_preset: null,
-          is_test: true,
-        },
-        {
-          environment_id: "cm9kddfcs01zqdy88ld9mmrli",
-          organization_id: "cm8zs78wb0002dy616dg75tv3",
-          project_id: "cm9kddfbz01zpdy88t9dstecu",
-          run_id: "cma45oli70002qrdy47w0j4n7",
-          event_name: "locked",
-          environment_type: null,
-          friendly_id: null,
-          attempt: 1,
-          engine: null,
-          status: "PENDING",
-          task_identifier: null,
-          queue: null,
-          schedule_id: null,
-          batch_id: null,
-          root_run_id: null,
-          parent_run_id: null,
-          depth: 0,
-          span_id: null,
-          trace_id: null,
-          idempotency_key: null,
-          event_time: new Date("2025-04-30 16:34:05.402").getTime(),
-          created_at: new Date("2025-04-30 16:34:04.312").getTime(),
-          updated_at: new Date("2025-04-30 16:34:05.378").getTime(),
-          started_at: null,
-          executed_at: new Date("2025-04-30 16:34:05.377").getTime(),
-          completed_at: null,
-          delay_until: null,
-          queued_at: null,
-          expired_at: null,
-          expiration_ttl: null,
-          usage_duration_ms: 0,
-          cost_in_cents: 0,
-          base_cost_in_cents: 0,
-          payload: null,
-          output: null,
-          error: null,
-          tags: [],
-          task_version: "20250430.3",
-          sdk_version: "4.0.0-v4-beta.7",
-          cli_version: "4.0.0-v4-beta.7",
-          machine_preset: "small-1x",
-          is_test: false,
-        },
-        {
-          environment_id: "cm9kddfcs01zqdy88ld9mmrli",
-          organization_id: "cm8zs78wb0002dy616dg75tv3",
-          project_id: "cm9kddfbz01zpdy88t9dstecu",
-          run_id: "cma45oli70002qrdy47w0j4n7",
-          event_name: "attempt_started",
-          environment_type: null,
-          friendly_id: null,
-          attempt: 1,
-          engine: null,
-          status: "EXECUTING",
-          task_identifier: null,
-          queue: null,
-          schedule_id: null,
-          batch_id: null,
-          root_run_id: null,
-          parent_run_id: null,
-          depth: 0,
-          span_id: null,
-          trace_id: null,
-          idempotency_key: null,
-          event_time: new Date("2025-04-30 16:34:08.129").getTime(),
-          created_at: new Date("2025-04-30 16:34:04.312").getTime(),
-          updated_at: new Date("2025-04-30 16:34:08.111").getTime(),
-          started_at: null,
-          executed_at: new Date("2025-04-30 16:34:08.112").getTime(),
-          completed_at: null,
-          delay_until: null,
-          queued_at: null,
-          expired_at: null,
-          expiration_ttl: null,
-          usage_duration_ms: 0,
-          cost_in_cents: 0,
-          base_cost_in_cents: 0,
-          payload: null,
-          output: null,
-          error: null,
-          tags: [],
-          task_version: null,
-          sdk_version: null,
-          cli_version: null,
-          machine_preset: null,
-          is_test: false,
-        },
-        {
-          environment_id: "cm9kddfcs01zqdy88ld9mmrli",
-          organization_id: "cm8zs78wb0002dy616dg75tv3",
-          project_id: "cm9kddfbz01zpdy88t9dstecu",
-          run_id: "cma45oli70002qrdy47w0j4n7",
-          event_name: "usage",
-          environment_type: null,
-          friendly_id: null,
-          attempt: 1,
-          engine: null,
-          status: "EXECUTING",
-          task_identifier: null,
-          queue: null,
-          schedule_id: null,
-          batch_id: null,
-          root_run_id: null,
-          parent_run_id: null,
-          depth: 0,
-          span_id: null,
-          trace_id: null,
-          idempotency_key: null,
-          event_time: new Date("2025-04-30 16:34:13.084").getTime(),
-          created_at: new Date("2025-04-30 16:34:04.000").getTime(),
-          updated_at: new Date("2025-04-30 16:34:13.000").getTime(),
-          started_at: null,
-          executed_at: null,
-          completed_at: null,
-          delay_until: null,
-          queued_at: null,
-          expired_at: null,
-          expiration_ttl: null,
-          usage_duration_ms: 2635,
-          cost_in_cents: 0.008893125,
-          base_cost_in_cents: 0,
-          payload: null,
-          output: null,
-          error: null,
-          tags: [],
-          task_version: null,
-          sdk_version: null,
-          cli_version: null,
-          machine_preset: null,
-          is_test: false,
-        },
-        {
-          environment_id: "cm9kddfcs01zqdy88ld9mmrli",
-          organization_id: "cm8zs78wb0002dy616dg75tv3",
-          project_id: "cm9kddfbz01zpdy88t9dstecu",
-          run_id: "cma45oli70002qrdy47w0j4n7",
-          event_name: "retry_scheduled",
-          environment_type: null,
-          friendly_id: null,
-          attempt: 2,
-          engine: null,
-          status: "RETRYING_AFTER_FAILURE",
-          task_identifier: null,
-          queue: null,
-          schedule_id: null,
-          batch_id: null,
-          root_run_id: null,
-          parent_run_id: null,
-          depth: 0,
-          span_id: null,
-          trace_id: null,
-          idempotency_key: null,
-          event_time: new Date("2025-04-30 16:34:13.748").getTime(),
-          created_at: new Date("2025-04-30 16:34:04.312").getTime(),
-          updated_at: new Date("2025-04-30 16:34:13.751").getTime(),
-          started_at: null,
-          executed_at: null,
-          completed_at: null,
-          delay_until: null,
-          queued_at: null,
-          expired_at: null,
-          expiration_ttl: null,
-          usage_duration_ms: 0,
-          cost_in_cents: 0,
-          base_cost_in_cents: 0,
-          payload: null,
-          output: null,
-          error: {
-            message: "Intentionally failing attempt 1",
-            name: "Error",
-            stackTrace:
-              "Error: Intentionally failing attempt 1\n    at run (file:///src/trigger/retry.ts:26:21)\n    at _tracer.startActiveSpan.attributes (file:///node_modules/.pnpm/@trigger.dev+core@4.0.0-v4-beta.7/node_modules/@trigger.dev/core/src/v3/workers/taskExecutor.ts:414:40)\n    at file:///node_modules/.pnpm/@trigger.dev+core@4.0.0-v4-beta.7/node_modules/@trigger.dev/core/src/v3/tracer.ts:141:24\n    at AsyncLocalStorage.run (node:async_hooks:346:14)\n    at AsyncLocalStorageContextManager.with (file:///node_modules/.pnpm/@opentelemetry+context-async-hooks@1.25.1_@opentelemetry+api@1.9.0/node_modules/@opentelemetry/context-async-hooks/src/AsyncLocalStorageContextManager.ts:40:36)\n    at ContextAPI2.with (file:///node_modules/.pnpm/@opentelemetry+api@1.9.0/node_modules/@opentelemetry/api/src/api/context.ts:77:42)\n    at Tracer.startActiveSpan (file:///node_modules/.pnpm/@opentelemetry+sdk-trace-base@1.25.1_@opentelemetry+api@1.9.0/node_modules/@opentelemetry/sdk-trace-base/src/Tracer.ts:241:24)\n    at TriggerTracer.startActiveSpan (file:///node_modules/.pnpm/@trigger.dev+core@4.0.0-v4-beta.7/node_modules/@trigger.dev/core/src/v3/tracer.ts:85:24)\n    at file:///node_modules/.pnpm/@trigger.dev+core@4.0.0-v4-beta.7/node_modules/@trigger.dev/core/src/v3/workers/taskExecutor.ts:409:33\n    at _RunTimelineMetricsAPI.measureMetric (file:///node_modules/.pnpm/@trigger.dev+core@4.0.0-v4-beta.7/node_modules/@trigger.dev/core/src/v3/runTimelineMetrics/index.ts:67:28)",
-            type: "BUILT_IN_ERROR",
-          },
-          tags: [],
-          task_version: null,
-          sdk_version: null,
-          cli_version: null,
-          machine_preset: null,
-          is_test: false,
-        },
-        {
-          environment_id: "cm9kddfcs01zqdy88ld9mmrli",
-          organization_id: "cm8zs78wb0002dy616dg75tv3",
-          project_id: "cm9kddfbz01zpdy88t9dstecu",
-          run_id: "cma45oli70002qrdy47w0j4n7",
-          event_name: "attempt_started",
-          environment_type: null,
-          friendly_id: null,
-          attempt: 2,
-          engine: null,
-          status: "EXECUTING",
-          task_identifier: null,
-          queue: null,
-          schedule_id: null,
-          batch_id: null,
-          root_run_id: null,
-          parent_run_id: null,
-          depth: 0,
-          span_id: null,
-          trace_id: null,
-          idempotency_key: null,
-          event_time: new Date("2025-04-30 16:34:13.846").getTime(),
-          created_at: new Date("2025-04-30 16:34:04.312").getTime(),
-          updated_at: new Date("2025-04-30 16:34:13.831").getTime(),
-          started_at: null,
-          executed_at: new Date("2025-04-30 16:34:08.112").getTime(),
-          completed_at: null,
-          delay_until: null,
-          queued_at: null,
-          expired_at: null,
-          expiration_ttl: null,
-          usage_duration_ms: 0,
-          cost_in_cents: 0,
-          base_cost_in_cents: 0,
-          payload: null,
-          output: null,
-          error: null,
-          tags: [],
-          task_version: null,
-          sdk_version: null,
-          cli_version: null,
-          machine_preset: null,
-          is_test: false,
-        },
-        {
-          environment_id: "cm9kddfcs01zqdy88ld9mmrli",
-          organization_id: "cm8zs78wb0002dy616dg75tv3",
-          project_id: "cm9kddfbz01zpdy88t9dstecu",
-          run_id: "cma45oli70002qrdy47w0j4n7",
-          event_name: "usage",
-          environment_type: null,
-          friendly_id: null,
-          attempt: 2,
-          engine: null,
-          status: "EXECUTING",
-          task_identifier: null,
-          queue: null,
-          schedule_id: null,
-          batch_id: null,
-          root_run_id: null,
-          parent_run_id: null,
-          depth: 0,
-          span_id: null,
-          trace_id: null,
-          idempotency_key: null,
-          event_time: new Date("2025-04-30 16:34:18.385").getTime(),
-          created_at: new Date("2025-04-30 16:34:04.000").getTime(),
-          updated_at: new Date("2025-04-30 16:34:18.000").getTime(),
-          started_at: null,
-          executed_at: null,
-          completed_at: null,
-          delay_until: null,
-          queued_at: null,
-          expired_at: null,
-          expiration_ttl: null,
-          usage_duration_ms: 5419,
-          cost_in_cents: 0.018289125,
-          base_cost_in_cents: 0,
-          payload: null,
-          output: null,
-          error: null,
-          tags: [],
-          task_version: null,
-          sdk_version: null,
-          cli_version: null,
-          machine_preset: null,
-          is_test: false,
-        },
-        {
-          environment_id: "cm9kddfcs01zqdy88ld9mmrli",
-          organization_id: "cm8zs78wb0002dy616dg75tv3",
-          project_id: "cm9kddfbz01zpdy88t9dstecu",
-          run_id: "cma45oli70002qrdy47w0j4n7",
-          event_name: "retry_scheduled",
-          environment_type: null,
-          friendly_id: null,
-          attempt: 3,
-          engine: null,
-          status: "RETRYING_AFTER_FAILURE",
-          task_identifier: null,
-          queue: null,
-          schedule_id: null,
-          batch_id: null,
-          root_run_id: null,
-          parent_run_id: null,
-          depth: 0,
-          span_id: null,
-          trace_id: null,
-          idempotency_key: null,
-          event_time: new Date("2025-04-30 16:34:18.832").getTime(),
-          created_at: new Date("2025-04-30 16:34:04.312").getTime(),
-          updated_at: new Date("2025-04-30 16:34:18.834").getTime(),
-          started_at: null,
-          executed_at: null,
-          completed_at: null,
-          delay_until: null,
-          queued_at: null,
-          expired_at: null,
-          expiration_ttl: null,
-          usage_duration_ms: 0,
-          cost_in_cents: 0,
-          base_cost_in_cents: 0,
-          payload: null,
-          output: null,
-          error: {
-            message: "Intentionally failing attempt 2",
-            name: "Error",
-            stackTrace:
-              "Error: Intentionally failing attempt 2\n    at run (file:///src/trigger/retry.ts:26:21)\n    at _tracer.startActiveSpan.attributes (file:///node_modules/.pnpm/@trigger.dev+core@4.0.0-v4-beta.7/node_modules/@trigger.dev/core/src/v3/workers/taskExecutor.ts:414:40)\n    at file:///node_modules/.pnpm/@trigger.dev+core@4.0.0-v4-beta.7/node_modules/@trigger.dev/core/src/v3/tracer.ts:141:24\n    at AsyncLocalStorage.run (node:async_hooks:346:14)\n    at AsyncLocalStorageContextManager.with (file:///node_modules/.pnpm/@opentelemetry+context-async-hooks@1.25.1_@opentelemetry+api@1.9.0/node_modules/@opentelemetry/context-async-hooks/src/AsyncLocalStorageContextManager.ts:40:36)\n    at ContextAPI2.with (file:///node_modules/.pnpm/@opentelemetry+api@1.9.0/node_modules/@opentelemetry/api/src/api/context.ts:77:42)\n    at Tracer.startActiveSpan (file:///node_modules/.pnpm/@opentelemetry+sdk-trace-base@1.25.1_@opentelemetry+api@1.9.0/node_modules/@opentelemetry/sdk-trace-base/src/Tracer.ts:241:24)\n    at TriggerTracer.startActiveSpan (file:///node_modules/.pnpm/@trigger.dev+core@4.0.0-v4-beta.7/node_modules/@trigger.dev/core/src/v3/tracer.ts:85:24)\n    at file:///node_modules/.pnpm/@trigger.dev+core@4.0.0-v4-beta.7/node_modules/@trigger.dev/core/src/v3/workers/taskExecutor.ts:409:33\n    at _RunTimelineMetricsAPI.measureMetric (file:///node_modules/.pnpm/@trigger.dev+core@4.0.0-v4-beta.7/node_modules/@trigger.dev/core/src/v3/runTimelineMetrics/index.ts:67:28)",
-            type: "BUILT_IN_ERROR",
-          },
-          tags: [],
-          task_version: null,
-          sdk_version: null,
-          cli_version: null,
-          machine_preset: null,
-          is_test: false,
-        },
-        {
-          environment_id: "cm9kddfcs01zqdy88ld9mmrli",
-          organization_id: "cm8zs78wb0002dy616dg75tv3",
-          project_id: "cm9kddfbz01zpdy88t9dstecu",
-          run_id: "cma45oli70002qrdy47w0j4n7",
-          event_name: "attempt_started",
-          environment_type: null,
-          friendly_id: null,
-          attempt: 3,
-          engine: null,
-          status: "EXECUTING",
-          task_identifier: null,
-          queue: null,
-          schedule_id: null,
-          batch_id: null,
-          root_run_id: null,
-          parent_run_id: null,
-          depth: 0,
-          span_id: null,
-          trace_id: null,
-          idempotency_key: null,
-          event_time: new Date("2025-04-30 16:34:18.946").getTime(),
-          created_at: new Date("2025-04-30 16:34:04.312").getTime(),
-          updated_at: new Date("2025-04-30 16:34:18.931").getTime(),
-          started_at: null,
-          executed_at: new Date("2025-04-30 16:34:08.112").getTime(),
-          completed_at: null,
-          delay_until: null,
-          queued_at: null,
-          expired_at: null,
-          expiration_ttl: null,
-          usage_duration_ms: 0,
-          cost_in_cents: 0,
-          base_cost_in_cents: 0,
-          payload: null,
-          output: null,
-          error: null,
-          tags: [],
-          task_version: null,
-          sdk_version: null,
-          cli_version: null,
-          machine_preset: null,
-          is_test: false,
-        },
-        {
-          environment_id: "cm9kddfcs01zqdy88ld9mmrli",
-          organization_id: "cm8zs78wb0002dy616dg75tv3",
-          project_id: "cm9kddfbz01zpdy88t9dstecu",
-          run_id: "cma45oli70002qrdy47w0j4n7",
-          event_name: "usage",
-          environment_type: null,
-          friendly_id: null,
-          attempt: 3,
-          engine: null,
-          status: "EXECUTING",
-          task_identifier: null,
-          queue: null,
-          schedule_id: null,
-          batch_id: null,
-          root_run_id: null,
-          parent_run_id: null,
-          depth: 0,
-          span_id: null,
-          trace_id: null,
-          idempotency_key: null,
-          event_time: new Date("2025-04-30 16:34:23.559").getTime(),
-          created_at: new Date("2025-04-30 16:34:04.000").getTime(),
-          updated_at: new Date("2025-04-30 16:34:23.000").getTime(),
-          started_at: null,
-          executed_at: null,
-          completed_at: null,
-          delay_until: null,
-          queued_at: null,
-          expired_at: null,
-          expiration_ttl: null,
-          usage_duration_ms: 8217,
-          cost_in_cents: 0.027732375,
-          base_cost_in_cents: 0,
-          payload: null,
-          output: null,
-          error: null,
-          tags: [],
-          task_version: null,
-          sdk_version: null,
-          cli_version: null,
-          machine_preset: null,
-          is_test: false,
-        },
-        {
-          environment_id: "cm9kddfcs01zqdy88ld9mmrli",
-          organization_id: "cm8zs78wb0002dy616dg75tv3",
-          project_id: "cm9kddfbz01zpdy88t9dstecu",
-          run_id: "cma45oli70002qrdy47w0j4n7",
-          event_name: "retry_scheduled",
-          environment_type: null,
-          friendly_id: null,
-          attempt: 4,
-          engine: null,
-          status: "RETRYING_AFTER_FAILURE",
-          task_identifier: null,
-          queue: null,
-          schedule_id: null,
-          batch_id: null,
-          root_run_id: null,
-          parent_run_id: null,
-          depth: 0,
-          span_id: null,
-          trace_id: null,
-          idempotency_key: null,
-          event_time: new Date("2025-04-30 16:34:24.045").getTime(),
-          created_at: new Date("2025-04-30 16:34:04.312").getTime(),
-          updated_at: new Date("2025-04-30 16:34:24.047").getTime(),
-          started_at: null,
-          executed_at: null,
-          completed_at: null,
-          delay_until: null,
-          queued_at: null,
-          expired_at: null,
-          expiration_ttl: null,
-          usage_duration_ms: 0,
-          cost_in_cents: 0,
-          base_cost_in_cents: 0,
-          payload: null,
-          output: null,
-          error: {
-            message: "Intentionally failing attempt 3",
-            name: "Error",
-            stackTrace:
-              "Error: Intentionally failing attempt 3\n    at run (file:///src/trigger/retry.ts:26:21)\n    at _tracer.startActiveSpan.attributes (file:///node_modules/.pnpm/@trigger.dev+core@4.0.0-v4-beta.7/node_modules/@trigger.dev/core/src/v3/workers/taskExecutor.ts:414:40)\n    at file:///node_modules/.pnpm/@trigger.dev+core@4.0.0-v4-beta.7/node_modules/@trigger.dev/core/src/v3/tracer.ts:141:24\n    at AsyncLocalStorage.run (node:async_hooks:346:14)\n    at AsyncLocalStorageContextManager.with (file:///node_modules/.pnpm/@opentelemetry+context-async-hooks@1.25.1_@opentelemetry+api@1.9.0/node_modules/@opentelemetry/context-async-hooks/src/AsyncLocalStorageContextManager.ts:40:36)\n    at ContextAPI2.with (file:///node_modules/.pnpm/@opentelemetry+api@1.9.0/node_modules/@opentelemetry/api/src/api/context.ts:77:42)\n    at Tracer.startActiveSpan (file:///node_modules/.pnpm/@opentelemetry+sdk-trace-base@1.25.1_@opentelemetry+api@1.9.0/node_modules/@opentelemetry/sdk-trace-base/src/Tracer.ts:241:24)\n    at TriggerTracer.startActiveSpan (file:///node_modules/.pnpm/@trigger.dev+core@4.0.0-v4-beta.7/node_modules/@trigger.dev/core/src/v3/tracer.ts:85:24)\n    at file:///node_modules/.pnpm/@trigger.dev+core@4.0.0-v4-beta.7/node_modules/@trigger.dev/core/src/v3/workers/taskExecutor.ts:409:33\n    at _RunTimelineMetricsAPI.measureMetric (file:///node_modules/.pnpm/@trigger.dev+core@4.0.0-v4-beta.7/node_modules/@trigger.dev/core/src/v3/runTimelineMetrics/index.ts:67:28)",
-            type: "BUILT_IN_ERROR",
-          },
-          tags: [],
-          task_version: null,
-          sdk_version: null,
-          cli_version: null,
-          machine_preset: null,
-          is_test: false,
-        },
-        {
-          environment_id: "cm9kddfcs01zqdy88ld9mmrli",
-          organization_id: "cm8zs78wb0002dy616dg75tv3",
-          project_id: "cm9kddfbz01zpdy88t9dstecu",
-          run_id: "cma45oli70002qrdy47w0j4n7",
-          event_name: "attempt_started",
-          environment_type: null,
-          friendly_id: null,
-          attempt: 4,
-          engine: null,
-          status: "EXECUTING",
-          task_identifier: null,
-          queue: null,
-          schedule_id: null,
-          batch_id: null,
-          root_run_id: null,
-          parent_run_id: null,
-          depth: 0,
-          span_id: null,
-          trace_id: null,
-          idempotency_key: null,
-          event_time: new Date("2025-04-30 16:34:24.135").getTime(),
-          created_at: new Date("2025-04-30 16:34:04.312").getTime(),
-          updated_at: new Date("2025-04-30 16:34:24.123").getTime(),
-          started_at: null,
-          executed_at: new Date("2025-04-30 16:34:08.112").getTime(),
-          completed_at: null,
-          delay_until: null,
-          queued_at: null,
-          expired_at: null,
-          expiration_ttl: null,
-          usage_duration_ms: 0,
-          cost_in_cents: 0,
-          base_cost_in_cents: 0,
-          payload: null,
-          output: null,
-          error: null,
-          tags: [],
-          task_version: null,
-          sdk_version: null,
-          cli_version: null,
-          machine_preset: null,
-          is_test: false,
-        },
-        {
-          environment_id: "cm9kddfcs01zqdy88ld9mmrli",
-          organization_id: "cm8zs78wb0002dy616dg75tv3",
-          project_id: "cm9kddfbz01zpdy88t9dstecu",
-          run_id: "cma45oli70002qrdy47w0j4n7",
-          event_name: "usage",
-          environment_type: null,
-          friendly_id: null,
-          attempt: 4,
-          engine: null,
-          status: "EXECUTING",
-          task_identifier: null,
-          queue: null,
-          schedule_id: null,
-          batch_id: null,
-          root_run_id: null,
-          parent_run_id: null,
-          depth: 0,
-          span_id: null,
-          trace_id: null,
-          idempotency_key: null,
-          event_time: new Date("2025-04-30 16:34:25.895").getTime(),
-          created_at: new Date("2025-04-30 16:34:04.000").getTime(),
-          updated_at: new Date("2025-04-30 16:34:25.000").getTime(),
-          started_at: null,
-          executed_at: null,
-          completed_at: null,
-          delay_until: null,
-          queued_at: null,
-          expired_at: null,
-          expiration_ttl: null,
-          usage_duration_ms: 8326,
-          cost_in_cents: 0.02810025,
-          base_cost_in_cents: 0,
-          payload: null,
-          output: null,
-          error: null,
-          tags: [],
-          task_version: null,
-          sdk_version: null,
-          cli_version: null,
-          machine_preset: null,
-          is_test: false,
-        },
-        {
-          environment_id: "cm9kddfcs01zqdy88ld9mmrli",
-          organization_id: "cm8zs78wb0002dy616dg75tv3",
-          project_id: "cm9kddfbz01zpdy88t9dstecu",
-          run_id: "cma45oli70002qrdy47w0j4n7",
-          event_name: "succeeded",
-          environment_type: null,
-          friendly_id: null,
-          attempt: 4,
-          engine: null,
-          status: "COMPLETED_SUCCESSFULLY",
-          task_identifier: null,
-          queue: null,
-          schedule_id: null,
-          batch_id: null,
-          root_run_id: null,
-          parent_run_id: null,
-          depth: 0,
-          span_id: null,
-          trace_id: null,
-          idempotency_key: null,
-          event_time: new Date("2025-04-30 16:34:26.139").getTime(),
-          created_at: new Date("2025-04-30 16:34:04.312").getTime(),
-          updated_at: new Date("2025-04-30 16:34:26.140").getTime(),
-          started_at: null,
-          executed_at: null,
-          completed_at: new Date("2025-04-30 16:34:26.139").getTime(),
-          delay_until: null,
-          queued_at: null,
-          expired_at: null,
-          expiration_ttl: null,
-          usage_duration_ms: 8326,
-          cost_in_cents: 0.02810025,
-          base_cost_in_cents: 0,
-          payload: null,
-          output: { attemptsTaken: "4" },
-          error: null,
-          tags: [],
-          task_version: null,
-          sdk_version: null,
-          cli_version: null,
-          machine_preset: null,
-          is_test: false,
-        },
-      ]);
-
-      expect(insertError).toBeNull();
-      expect(insertResult).toEqual(expect.objectContaining({ executed: true }));
-      expect(insertResult?.summary?.written_rows).toEqual("15");
-
-      const query = client.query({
-        name: "query-run-events",
-        query: "SELECT * FROM trigger_dev.raw_run_events_v1",
-        schema: z.object({
-          environment_id: z.string(),
-          run_id: z.string(),
-        }),
-        params: z.object({
-          run_id: z.string(),
-        }),
-      });
-
-      const [queryError, result] = await query({ run_id: "run_1234" });
-
-      expect(queryError).toBeNull();
-      expect(result).toEqual(
-        expect.arrayContaining([
-          expect.objectContaining({
-            environment_id: "cm9kddfcs01zqdy88ld9mmrli",
-            run_id: "cma45oli70002qrdy47w0j4n7",
-          }),
-        ])
-      );
-
-      // Query the materialized view to check the final state
-      const latestQuery = client.query({
-        name: "query-run-latest",
-        query: "SELECT * FROM trigger_dev.run_latest_v1 FINAL",
-        schema: z.any(),
-      });
-
-      const [latestQueryError, latestResult] = await latestQuery({});
-
-      console.log(latestResult);
-
-      expect(latestQueryError).toBeNull();
-      expect(latestResult).not.toBeNull();
-      if (!latestResult) throw new Error("Expected latestResult to not be null");
-      expect(latestResult).toHaveLength(1);
-      expect(latestResult[0]).toEqual(
-        expect.objectContaining({
-          environment_id: "cm9kddfcs01zqdy88ld9mmrli",
-          run_id: "cma45oli70002qrdy47w0j4n7",
-          status: "COMPLETED_SUCCESSFULLY",
-          attempt: 4,
-          task_identifier: "retry-task",
-          task_version: "20250430.3",
-          sdk_version: "4.0.0-v4-beta.7",
-          cli_version: "4.0.0-v4-beta.7",
-          machine_preset: "small-1x",
-          usage_duration_ms: 8326,
-          cost_in_cents: 0.02810025,
-          base_cost_in_cents: 0,
-          payload: { failCount: "3" },
-          output: { attemptsTaken: "4" },
-          error: null,
-          completed_at: "2025-04-30 15:34:26.139",
-          is_test: 0,
-          environment_type: "PRODUCTION",
-          friendly_id: "run_cma45oli70002qrdy47w0j4n7",
-          queue: "task/retry-task",
-          schedule_id: null,
-          batch_id: null,
-          root_run_id: null,
-          depth: 0,
-          started_at: null,
-          delay_until: null,
-          expiration_ttl: null,
-          expired_at: null,
-          span_id: "538677637f937f54",
-          idempotency_key: null,
-          tags: [],
-          created_at: "2025-04-30 15:34:04.312",
-          _version: "2025-04-30 15:34:26.139",
-          last_event_time: "2025-04-30 15:34:26.139",
-          updated_at: "2025-04-30 15:34:26.140",
-          engine: "V2",
-        })
-      );
-    }
-  );
-});
